requirements.txt
torch>=1.9.0
numpy>=1.21.0
pyyaml>=5.4.0
matplotlib>=3.5.0
pandas>=1.3.0
